Add Rock Paper Scissors program
This is a program that asks the user to write either "Rock", "Paper", or "Scissors". The program will randomly select its own selection and depending on what the player picks, as well as the computer, the outcome will show whether the player won, the computer won, or it was a tie. It will then loop through and ask the user for their input again.