Add Calculator
This is a simple calculator I made using Python, it asks the user for what operation to use ( 1- Addition, 2- Subtraction, 3- Multiplication, 4- Division). It will then ask for the first number, along with the second number and will then display the output. After the output, once a key is hit the application will close.